qualcommax: ipq807x: ZyXEL NBG7815: DualBoot Support (1)
In order to change the bootconfig it is necessary to modify the DTS file.
mtd2 and mtd3 have to be writeable.

ZyXEL is writing both bootpartitions mtd2/0:bootconfig and mtd3/0:bootconfig1
every time during firmware update.

As we've installed OpenWrt accordingly what is written in the Wiki we only
changed the bootconfig for "rootfs" (seek 168). Which is basically wrong.

Flashing ZyXEL firmware via tftp is fixing this order. Because they
are writing the order for kernel and wififw too. Always accordingly what
rootfs is giving them and regardless how often you flash the firmware over
the device through either gui or tftp.
They are changing always this three partitions. The other partitions are not
touched and managed by the bootloader/zloader for failsafe boot.

cat /proc/boot_info/*/upgradepartition

0:APDP_1
0:APPSBL_1
0:CDT_1
0:DEVCFG_1
0:HLOS
0:QSEE_1
0:RPM_1
0:WIFIFW
rootfs

cat /proc/boot_info/*/primaryboot

0
0
0
0
1
0
0
1
1

1 is here kernel, wififw and rootfs. The 1 is reflected accordingly what's
written within upgradepartition.

Within the partition /dev/mtd[2,3] the order is not alphabetical like the
output above.

0:QSEE
0:DEVCFG
0:APDP
0:RPM
0:CDT
0:APPSBL
0:HLOS
rootfs
0:WIFIFW

--> seek 148=kernel, 168=rootfs, 188=wififw

So for upgrading OpenWrt writing the opposite partition we are modifying
this three parts of the bootconfig like ZyXEL is doing it.